WIP: attempt to fix LSP completion issues
So far this fixes the `(` being inserted in https://github.com/godotengine/godot-vscode-plugin/issues/142
and also the extra quotes, but only if completion is
trigged by the quote key.

It does not yet fix https://github.com/godotengine/godot-vscode-plugin/issues/184
because the third param to `connect()` actually is a method.
We need more context to be able to determine whether or not to add parentheses.